Exposed several glm matrix helper functions.  Randomly segfaults.
I'm pretty sure that the segfaulting has to do with garbage
collection.  I'm probably doing reference counting wrong.  I'm
probably doing it wrong in a bunch of places.

It seems that you just need to do enough of the type operation things
for it to buy the farm.  Segfault occurs when trying to init a new
object with one of the custom types.